# backend/docs/swagger.yaml
definitions:
...
  domain.CareGuide:
    properties:
      created_at:
        type: string
      id:
        type: integer
      instructions:
        type: object
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  domain.Category:
    properties:
      cover_image:
        $ref: '#/definitions/domain.Media'
      cover_image_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
      updated_at:
        type: string
    type: object
  domain.CreateCareGuideRequest:
    properties:
      instructions:
        type: object
      title:
        type: string
    required:
    - title
    type: object
  domain.CreateCategoryRequest:
    properties:
      cover_image_id:
        type: integer
      name:
        type: string
      slug:
        type: string
    required:
    - name
    - slug
    type: object
...
  domain.Media:
    properties:
      alt:
        type: string
      created_at:
        type: string
      filename:
        type: string
      filesize:
        type: integer
      height:
        type: integer
      id:
        type: integer
      mime_type:
        type: string
      updated_at:
        type: string
      url:
        type: string
      width:
        type: integer
    type: object
...
  domain.UpdateCareGuideRequest:
    properties:
      instructions:
        type: object
      title:
        type: string
    required:
    - title
    type: object
  domain.UpdateCategoryRequest:
    properties:
      cover_image_id:
        type: integer
      name:
        type: string
      slug:
        type: string
    required:
    - name
    - slug
    type: object
...
  response.ListMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  response.ListResponse-domain_CareGuide:
    properties:
      code:
        type: string
      data:
        items:
          $ref: '#/definitions/domain.CareGuide'
        type: array
      message:
        type: string
      meta:
        $ref: '#/definitions/response.ListMeta'
      success:
        type: boolean
    type: object
  response.ListResponse-domain_Category:
    properties:
      code:
        type: string
      data:
        items:
          $ref: '#/definitions/domain.Category'
        type: array
      message:
        type: string
      meta:
        $ref: '#/definitions/response.ListMeta'
      success:
        type: boolean
    type: object
...
  response.Response-any:
    properties:
      code:
        type: string
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  response.Response-domain_CareGuide:
    properties:
      code:
        type: string
      data:
        $ref: '#/definitions/domain.CareGuide'
      message:
        type: string
      success:
        type: boolean
    type: object
  response.Response-domain_Category:
    properties:
      code:
        type: string
      data:
        $ref: '#/definitions/domain.Category'
      message:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact: {}
  description: Customer auth, product catalog, and admin CMS API for the Diagonals
    website backend.
  title: Diagonals API
  version: "1.0"
paths:
...
  /api/v1/care-guides:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all care guides
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ListResponse-domain_CareGuide'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: Get all care guides
      tags:
      - CareGuides
    post:
      consumes:
      - application/json
      description: Create a new care guide
      parameters:
      - description: Care Guide payload
        in: body
        name: care_guide
        required: true
        schema:
          $ref: '#/definitions/domain.CreateCareGuideRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response-domain_CareGuide'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: Create care guide
      tags:
      - CareGuides
  /api/v1/care-guides/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a care guide by ID
      parameters:
      - description: Care Guide ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: Delete care guide
      tags:
      - CareGuides
    get:
      consumes:
      - application/json
      description: Retrieve a single care guide by ID
      parameters:
      - description: Care Guide ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response-domain_CareGuide'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: Get care guide by ID
      tags:
      - CareGuides
    put:
      consumes:
      - application/json
      description: Update an existing care guide
      parameters:
      - description: Care Guide ID
        in: path
        name: id
        required: true
        type: integer
      - description: Care Guide payload
        in: body
        name: care_guide
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateCareGuideRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response-domain_CareGuide'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: Update care guide
      tags:
      - CareGuides
...
  /api/v1/categories:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all product categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ListResponse-domain_Category'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: Get all categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Create a new product category
      parameters:
      - description: Category payload
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/domain.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response-domain_Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response-any'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: Create category
      tags:
      - Categories
  /api/v1/categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: Delete category
      tags:
      - Categories
    get:
      consumes:
      - application/json
      description: Retrieve a single category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response-domain_Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: Get category by ID
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Update an existing product category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Category payload
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response-domain_Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: Update category
      tags:
      - Categories
  /api/v1/categories/slug/{slug}:
    get:
      consumes:
      - application/json
      description: Retrieve a single category by slug
      parameters:
      - description: Category slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response-domain_Category'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: Get category by slug
      tags:
      - Categories
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"